refactor(crypto): hoist attachment crypto + sealed-payload parser into @threa/crypto
Slice C foundation. The AttachmentRef type, the per-file decrypt primitive, and
the sealed-payload parser/serializer were frontend-only — the enclave needs them
to read encrypted attachments, and a parallel impl would violate INV-35. Move
the canonical definitions into @threa/crypto (new attachment.ts + sealed-payload.ts);
the frontend modules now re-export them so no call site changes. No behavior
change — crypto (33) + frontend attachment/envelope (20) suites green.
