+++ apps/frontend/src/lib/crypto/attachment-crypto.ts
@@ -1,39 +1,16 @@
 import {
-  base64ToBytes,
+  ATTACHMENT_AAD,
+  ATTACHMENT_KEY_GENERATION,
   bytesToBase64,
   generateStreamKey,
-  openMessage,
   sealMessage,
-  STREAM_ENVELOPE_VERSION,
-  utf8Encode,
+  type AttachmentRef,
 } from "@threa/crypto"
 
-/**
- * One end-to-end-encrypted attachment, carried inside the SSK-sealed message
- * payload (its `attachmentRefs`) and never on the wire in clear. `key`/`iv`
- * open the opaque S3 ciphertext; `filename`/`mimeType`/`sizeBytes` are the real
- * values the server's placeholder row deliberately hides. See
- * `.claude/plans/e2e-attachments.md`.
- */
-export interface AttachmentRef {
-  attachmentId: string
-  /** Base64 of the 32-byte single-use AES-256-GCM key for this attachment. */
-  key: string
-  /** Base64 12-byte GCM IV (from the seal envelope). */
-  iv: string
-  filename: string
-  mimeType: string
-  sizeBytes: number
-}
-
-// Domain-separation label bound as GCM AAD. The per-attachment key is random
-// and used exactly once, so relocation/confusion attacks gain nothing and the
-// AAD's only job is to satisfy the AEAD interface and pin the ciphertext to
-// this scheme. It carries no secret and is reconstructed verbatim by the
-// decrypt path (Slice B2).
-const ATTACHMENT_AAD = utf8Encode("threa-attachment-v1")
-/** Single-key scheme: attachment keys are per-file, never rotated. */
-const ATTACHMENT_KEY_GENERATION = 0
+// The `AttachmentRef` type and the `decryptAttachmentBytes` open primitive are
+// canonical in @threa/crypto so the enclave shares them (INV-35). Re-export them
+// here so existing call sites that import from this module keep working.
+export { decryptAttachmentBytes, type AttachmentRef } from "@threa/crypto"
 
 export interface EncryptedAttachment {
   /** Ciphertext bytes to upload as the opaque file body (a valid `BlobPart`). */
@@ -59,30 +36,6 @@ export async function encryptAttachmentBytes(plaintext: Uint8Array): Promise<Enc
     aad: ATTACHMENT_AAD,
   })
   return { ciphertext, key: bytesToBase64(key), iv: envelope.iv }
-}
-
-/**
- * Decrypt the opaque S3 ciphertext of an E2E attachment back to its bytes, using
- * the `key`/`iv` carried in the message's `attachmentRef`. Inverse of
- * `encryptAttachmentBytes` — reconstructs the same single-key envelope (gen 0,
- * the domain-separation AAD) and opens it. Throws if the key/iv don't match or
- * the bytes were tampered (AES-GCM tag check).
- */
-export async function decryptAttachmentBytes(input: {
-  ciphertext: Uint8Array
-  key: string
-  iv: string
-}): Promise<Uint8Array<ArrayBuffer>> {
-  return openMessage({
-    key: base64ToBytes(input.key),
-    envelope: {
-      v: STREAM_ENVELOPE_VERSION,
-      keyGeneration: ATTACHMENT_KEY_GENERATION,
-      iv: input.iv,
-      aad: bytesToBase64(ATTACHMENT_AAD),
-    },
-    ciphertext: input.ciphertext,
-  })
 }
 
 // In-memory bridge from upload time (where the per-attachment key/iv are
