+++ apps/frontend/src/lib/crypto/message-envelope.ts
@@ -8,13 +8,20 @@ import {
   decryptPayloadAsString,
   ENVELOPE_VERSION,
   openMessageAsString,
+  parseSealedPayload,
   sealMessage,
+  serializeSealedPayload,
   STREAM_ENVELOPE_VERSION,
+  type AttachmentRef,
   type Envelope,
   type StreamEnvelope,
 } from "@threa/crypto"
 import { resolveStreamKey } from "./stream-key-cache"
-import type { AttachmentRef } from "./attachment-crypto"
+
+// Parsing/serializing the sealed payload (and the AttachmentRef shape) is shared
+// crypto so the enclave strips the same wrapper (INV-35). Re-export the parser
+// and its result type for callers that import them from this module.
+export { parseSealedPayload, type ParsedSealedPayload } from "@threa/crypto"
 
 // The placeholder text the backend stores in `contentMarkdown` / `contentJson`
 // for E2E messages is the single source of truth in @threa/types so the
@@ -37,83 +44,6 @@ export interface SealStreamMessageInput {
    * wire — the server only holds opaque bytes and a placeholder row.
    */
   attachmentRefs?: AttachmentRef[]
-}
-
-/**
- * The structured E2E message payload. Messages with no attachments seal the
- * bare markdown string (byte-identical to every E2E message already written),
- * so this wrapper only appears once attachments ride along. The `__e2ePayload`
- * marker + version lets the decrypt path tell a wrapper apart from a user's
- * markdown that merely happens to start with `{` — the same structurally-
- * disjoint discrimination the backend's envelope union uses.
- */
-interface E2eSealedPayload {
-  __e2ePayload: typeof E2E_PAYLOAD_VERSION
-  contentMarkdown: string
-  attachmentRefs: AttachmentRef[]
-}
-
-const E2E_PAYLOAD_VERSION = 1
-
-/** Build the bytes to seal: bare markdown, or the wrapper when refs ride along. */
-function serializeSealedPayload(contentMarkdown: string, attachmentRefs?: AttachmentRef[]): string {
-  if (!attachmentRefs || attachmentRefs.length === 0) return contentMarkdown
-  return JSON.stringify({
-    __e2ePayload: E2E_PAYLOAD_VERSION,
-    contentMarkdown,
-    attachmentRefs,
-  } satisfies E2eSealedPayload)
-}
-
-export interface ParsedSealedPayload {
-  contentMarkdown: string
-  attachmentRefs: AttachmentRef[]
-}
-
-/**
- * Validate one decrypted `attachmentRefs` element before it reaches the viewer.
- * The refs are decrypted text we authored, but a malformed or mixed-version
- * payload could carry objects missing `key`/`iv`/`filename`/etc. — those must
- * not flow to the timeline (it would try to fetch/decrypt with `undefined`).
- */
-function isAttachmentRef(value: unknown): value is AttachmentRef {
-  if (typeof value !== "object" || value === null) return false
-  const r = value as Record<string, unknown>
-  return (
-    typeof r.attachmentId === "string" &&
-    typeof r.key === "string" &&
-    typeof r.iv === "string" &&
-    typeof r.filename === "string" &&
-    typeof r.mimeType === "string" &&
-    typeof r.sizeBytes === "number"
-  )
-}
-
-/**
- * Inverse of `serializeSealedPayload`. A decrypted string is either the bare
- * markdown body (the legacy/no-attachment shape) or the versioned wrapper;
- * anything that doesn't parse as our wrapper is treated as raw markdown so old
- * messages keep opening unchanged.
- */
-export function parseSealedPayload(raw: string): ParsedSealedPayload {
-  if (raw.startsWith("{")) {
-    try {
-      const parsed = JSON.parse(raw) as Partial<E2eSealedPayload>
-      if (parsed.__e2ePayload === E2E_PAYLOAD_VERSION && typeof parsed.contentMarkdown === "string") {
-        // attachmentRefs comes from decrypted text we authored, but guard the
-        // shape anyway so a malformed wrapper degrades to "no attachments" and
-        // any malformed element is dropped, rather than handing junk to the
-        // timeline.
-        const attachmentRefs = Array.isArray(parsed.attachmentRefs)
-          ? parsed.attachmentRefs.filter(isAttachmentRef)
-          : []
-        return { contentMarkdown: parsed.contentMarkdown, attachmentRefs }
-      }
-    } catch {
-      // Not our wrapper — fall through and treat the whole string as markdown.
-    }
-  }
-  return { contentMarkdown: raw, attachmentRefs: [] }
 }
 
 export interface SealStreamMessageResult {
