+++ packages/crypto/src/index.ts
@@ -44,3 +44,19 @@ export {
   type WrapStreamKeyInput,
   type UnwrapStreamKeyInput,
 } from "./stream-key"
+
+export {
+  ATTACHMENT_AAD,
+  ATTACHMENT_KEY_GENERATION,
+  decryptAttachmentBytes,
+  type AttachmentRef,
+} from "./attachment"
+
+export {
+  E2E_PAYLOAD_VERSION,
+  serializeSealedPayload,
+  parseSealedPayload,
+  isAttachmentRef,
+  type E2eSealedPayload,
+  type ParsedSealedPayload,
+} from "./sealed-payload"
